Refactored Constants, Parsers (args, ci, and trace files), TimeUtils, and FileWriters out of the main CarbonFootprint module
