+++ src/Constants.py
@@ -1,0 +1,19 @@
+# Default Values
+DEFAULT = "default"
+FILE = "csv"
+DELIMITER = ","
+TRACE = "trace"
+CI = "ci"
+PUE = "pue"
+CORE_POWER_DRAW = "core-power-draw"
+MEMORY_COEFFICIENT = "memory-coefficient"
+MIN_WATTS = "min-watts"
+MAX_WATTS = "max-watts"
+GA = "GA"
+CCF = "CCF"
+BOTH = "BOTH"
+DEFAULT_PUE_VALUE = 1.0  # Disregard PUE if 1.0
+DEFAULT_MEMORY_POWER_DRAW = 0.392  # W/GB
+RESERVED_MEMORY = "reserved-memory"
+NUM_OF_NODES = "num-of-nodes"
+TASK_FLAG = True
