+++ src/scripts/CarbonFootprint.py
@@ -1,197 +1,14 @@
-from src.models.TraceRecord import TraceRecord
-from src.models.CarbonRecord import CarbonRecord, HEADERS
 import sys
-import datetime as time
-import copy
-import os
 
-
-# Default Values
-DEFAULT = "default"
-FILE = "csv"
-DELIMITER = ","
-TRACE = "trace"
-CI = "ci"
-PUE = "pue"
-CORE_POWER_DRAW = "core-power-draw"
-MEMORY_COEFFICIENT = "memory-coefficient"
-MIN_WATTS = "min-watts"
-MAX_WATTS = "max-watts"
-GA = "GA"
-CCF = "CCF"
-BOTH = "BOTH"
-DEFAULT_PUE_VALUE = 1.0  # Disregard PUE if 1.0
-DEFAULT_MEMORY_POWER_DRAW = 0.392  # W/GB
-RESERVED_MEMORY = "reserved-memory"
-NUM_OF_NODES = "num-of-nodes"
-TASK_FLAG = True
-
+from src.models.CarbonRecord import CarbonRecord
+from src.Constants import *
+from src.utils.Parsers import parse_arguments, parse_ci_intervals
+from src.utils.TimeUtils import extract_tasks_by_hour, to_timestamp
+from src.utils.FileWriters import write_trace_file, write_summary_file
 
 # Functions
 def linear_power_model(cpu_usage, min_watts, max_watts):
     return min_watts + cpu_usage * (max_watts - min_watts)
-
-
-# todo: timezone conversion for non-utc times
-
-
-def to_timestamp(ms):
-    return time.datetime.fromtimestamp(float(ms) / 1000.0, tz=time.timezone.utc)
-
-
-def get_ci_file_data(filename):
-    with open(filename, 'r') as file:
-        raw = file.readlines()
-        header = [val.strip() for val in raw[0].split(",")]
-        data = raw[1:]
-
-    return (header, data)
-
-
-def parse_ci_intervals(filename):
-    (header, data) = get_ci_file_data(filename)
-
-    date_i = header.index("date")
-    start_i = header.index("start")
-    value_i = header.index("actual")
-
-    ci_map = {}
-
-    for row in data:
-        parts = row.split(",")
-        date = parts[date_i]
-        month_day = '/'.join([val.zfill(2) for val in date.split('-')[-2:]])
-        key = month_day + '-' + parts[start_i]
-        value = float(parts[value_i])
-        ci_map[key] = value
-
-    return ci_map
-
-
-def parse_trace_file(filepath):
-    with open(filepath, 'r') as file:
-        lines = [line.rstrip() for line in file]
-
-    header = lines[0]
-    records = []
-
-    for line in lines[1:]:
-        trace_record = TraceRecord(header, line, DELIMITER)
-        records.append(trace_record)
-
-    return records
-
-
-def print_usage_exit():
-    usage = "Ichnos: python -m src.scripts.CarbonFootprint <trace-name> <ci-value|ci-file-name> <min-watts> <max-watts> <? pue=1.0> <? memory-coeff=0.392>"
-    print(usage)
-    exit(-1)
-
-
-def get_carbon_record(record: TraceRecord):
-    return record.make_carbon_record()
-
-
-def get_tasks_by_hour_with_overhead(start_hour, end_hour, tasks):
-    tasks_by_hour = {}
-    overheads = []
-    runtimes = []
-
-    step = 60 * 60 * 1000  # 60 minutes in ms
-    i = start_hour - step  # start an hour before to be safe
-
-    while i <= end_hour:
-        data = [] 
-        hour_overhead = 0
-
-        for task in tasks: 
-            start = int(task.get_start())
-            complete = int(task.get_complete())
-            # full task is within this hour
-            if start >= i and complete <= i + step:
-                data.append(task)
-                runtimes.append(complete - start)
-            # task ends within this hour (but starts in a previous hour)
-            elif complete > i and complete < i + step and start < i:
-                # add task from start of this hour until end of hour
-                partial_task = copy.deepcopy(task)
-                partial_task.set_start(i)
-                partial_task.set_realtime(complete - i)
-                data.append(partial_task)
-                runtimes.append(complete - i)
-            # task starts within this hour (but ends in a later hour) -- OVERHEAD
-            elif start > i and start < i + step and complete > i + step: 
-                # add task from start to end of this hour
-                partial_task = copy.deepcopy(task)
-                partial_task.set_complete(i + step)
-                partial_task.set_realtime(i + step - start)
-                data.append(partial_task)
-                if (i + step - start) > hour_overhead:  # get the overhead for the longest task that starts now but ends later
-                    hour_overhead = i + step - start
-                runtimes.append(i + step - start)
-            # task starts before hour and ends after this hour
-            elif start < i and complete > i + step:
-                partial_task = copy.deepcopy(task)
-                partial_task.set_start(i)
-                partial_task.set_complete(i + step)
-                partial_task.set_realtime(step)
-                data.append(partial_task)
-                runtimes.append(step)
-
-        tasks_by_hour[i] = data
-        overheads.append(hour_overhead)
-        i += step
-
-    # task_overall_runtime = sum(runtimes)
-
-    return (tasks_by_hour, overheads)
-
-
-def to_closest_hour_ms(original):
-    ts = to_timestamp(original)
-
-    if ts.minute >= 30:
-        if ts.hour + 1 == 24:
-            # ts = ts.replace(hour=0, minute=0, second=0, microsecond=0, day=ts.day+1)
-            ts = ts + time.timedelta(days=1)
-            ts = ts.replace(hour=0, minute=0, second=0, microsecond=0)
-        else:
-            ts = ts.replace(second=0, microsecond=0, minute=0, hour=ts.hour+1)
-    else:
-        ts = ts.replace(second=0, microsecond=0, minute=0)
-
-    return int(ts.timestamp() * 1000)  # closest hour in ms
-
-
-def get_tasks_by_hour(tasks):
-    starts = []
-    ends = []
-
-    for task in tasks:
-        starts.append(int(task.get_start()))
-        ends.append(int(task.get_complete()))
-
-    earliest = min(starts)
-    latest = max(ends)
-    earliest_hh = to_closest_hour_ms(earliest)  
-    latest_hh = to_closest_hour_ms(latest)
-
-    return get_tasks_by_hour_with_overhead(earliest_hh, latest_hh, tasks)
-
-
-def extract_tasks_by_hour(filename):
-    if len(filename.split(".")) > 1:
-        filename = filename.split(".")[-2]
-
-    records = parse_trace_file(f"data/trace/{filename}.{FILE}")
-    data_records = []
-
-    for record in records:
-        data = get_carbon_record(record)
-        data_records.append(data)
-
-    return get_tasks_by_hour(data_records)
-
 
 # Estimate Energy Consumption using CCF Methodology
 def estimate_task_energy_consumption_ccf(task: CarbonRecord, min_watts, max_watts, memory_coefficient):
@@ -263,80 +80,6 @@ def calculate_carbon_footprint_ccf(tasks_by_hour, ci, pue: float, min_watts, max
                 records.append(task)
 
     return ((total_energy, total_energy_pue, total_memory_energy, total_memory_energy_pue, total_carbon_emissions, node_memory_used), records)
-
-
-def get_hours(arr):
-    hours = []
-    prev = arr[0]
-    i = 1
-
-    while i < len(arr):
-        if not (prev + 1 == arr[i]):  # if not consecutive, workflow halts and resumes
-            hours.append(i - 1)  # add the overhead for the previous hour which will not finish by this hour
-        prev = arr[i]
-        i += 1
-
-    return hours
-
-
-def check_if_float(value):
-    return value.replace('.', '').isnumeric()
-
-
-def parse_arguments(args):
-    if len(args) != 4 and len(args) != 6 and len(args) != 8:
-        print_usage_exit()
-
-    arguments = {}
-    arguments[TRACE] = args[0]
-
-    if check_if_float(args[1]):
-        arguments[CI] = float(args[1])
-    else:
-        arguments[CI] = args[1]
-
-    arguments[MIN_WATTS] = float(args[2])
-    arguments[MAX_WATTS] = float(args[3])
-
-    if len(args) == 6:
-        arguments[PUE] = float(args[4])
-        arguments[MEMORY_COEFFICIENT] = float(args[5])
-    elif len(args) == 8:
-        arguments[PUE] = float(args[4])
-        arguments[MEMORY_COEFFICIENT] = float(args[5])
-        arguments[RESERVED_MEMORY] = float(args[6])
-        arguments[NUM_OF_NODES] = int(args[7])
-    else:
-        arguments[PUE] = DEFAULT_PUE_VALUE
-        arguments[MEMORY_COEFFICIENT] = DEFAULT_MEMORY_POWER_DRAW
-
-    return arguments
-
-
-def write_trace_file(folder, trace_file, records):
-    # Create the folder if it doesn't exist
-    if not os.path.exists(folder):
-        os.makedirs(folder)
-
-    output_file_name = f"{folder}/{trace_file}-trace.csv"
-
-    with open(output_file_name, "w") as file:
-        file.write(f"{HEADERS}\n")
-
-        for record in records:
-            file.write(f"{record}\n")
-
-
-def write_summary_file(folder, trace_file, content):
-    # Create the folder if it doesn't exist
-    if not os.path.exists(folder):
-        os.makedirs(folder)
-
-    output_file_name = f"{folder}/{trace_file}-summary.txt"
-
-    with open(output_file_name, "w") as file:
-        file.write(content)
-
 
 def main(arguments):
     # Data
